requirements.txt
beautifulsoup4==4.12.2
celery==5.3.4
Django==4.2.7
django-celery-beat==2.5.0
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
lxml==4.9.3
schedule==1.2.1
